+++ e2e/index.d.ts
@@ -1,0 +1,11 @@
+declare module '*prisma/generated/mongodb' {
+  class PrismaClient {
+    $runCommandRaw(command: unknown): any;
+  }
+}
+
+declare module '*prisma/generated/mysql' {
+  class PrismaClient {
+    $queryRawUnsafe(query: string): any;
+  }
+}