+++ lib/health-indicator/index.ts
@@ -7,6 +7,7 @@ export * from './database/mongoose.health';
 export * from './database/typeorm.health';
 export * from './database/mikro-orm.health';
 export * from './database/sequelize.health';
+export * from './database/prisma.health';
 export * from './microservice/microservice.health';
 export * from './microservice/grpc.health';
 export * from './disk';