+++ sample/012-prisma-app/src/health/health.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { PrismaModule } from 'src/prisma/prisma.module';
+import { HealthController } from './health.controller';
+import { TerminusModule } from '@nestjs/terminus';
+
+@Module({
+  imports: [PrismaModule, TerminusModule],
+  controllers: [HealthController],
+})
+export class HealthModule {}